requirements.txt
Django==5.0.9
djangorestframework==3.15.2
celery==5.4.0
redis==5.0.8
drf-yasg==1.21.7
django-ipware==7.0.1
django-ip-geolocation==1.6.1
django-ratelimit==4.1.0
gunicorn==23.0.0
whitenoise==6.7.0
python-decouple==3.8
psycopg2-binary==2.9.11
django-celery-beat==2.6.0
django-celery-results==2.5.1
ratelimit==4.1.0
decorator==5.1.1
